pyproject.toml
[project]
name = "ai_service"
version = "0.1.0"
description = "Compare two revisions of DOCX/PDF and analyze changes with multi-agent LLM + web search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.1",
  "httpx>=0.27.0",
  "tenacity>=8.2.3",
  "tika>=2.6.0",
  "python-docx>=1.1.0",
  "pdfplumber>=0.11.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[project.scripts]
compare-ai = "app.cli:main"

[tool.ruff]
line-length = 100

